+++ gui/src/app.rs
@@ -1,6 +1,7 @@
 use super::{AppEvent, GameGoals, GameViewState, SecCardGameApp};
 use crate::init_view::state::InitViewState;
-use egui::Context;
+use eframe::Frame;
+use egui::{Context, Ui};
 use game_lib::cards::game_variants::scenario::Scenario;
 use game_lib::world::deck::Deck;
 use game_lib::world::game::{Game, GameInitSettings};
@@ -48,14 +49,18 @@ impl SecCardGameApp {
 }
 
 impl eframe::App for SecCardGameApp {
-    /// Called each time the UI needs repainting, which may be many times per second.
-    fn update(&mut self, ctx: &Context, _frame: &mut eframe::Frame) {
+    /// Called before UI is drawn and also if window is hidden. Allows clear split for logic and drawing functions.
+    fn logic(&mut self, _ctx: &Context, _frame: &mut Frame) {
         self.handle_app_event();
+        self.active_view.logic()
+    }
 
-        self.create_menu_bar(ctx);
+    /// Called each time the UI needs repainting, which may be many times per second.
+    fn ui(&mut self, ui: &mut Ui, _frame: &mut Frame) {
+        self.create_menu_bar(ui);
 
         let mut event_publisher = |event| self.last_event = Some(event);
-        self.active_view.draw_ui(&mut event_publisher, ctx);
+        self.active_view.ui(&mut event_publisher, ui);
     }
 }
 
@@ -80,9 +85,9 @@ impl SecCardGameApp {
         };
     }
 
-    fn create_menu_bar(&mut self, ctx: &Context) {
-        egui::TopBottomPanel::top("top_panel").show(ctx, |ui| {
-            egui::menu::bar(ui, |ui| {
+    fn create_menu_bar(&mut self, ui: &mut Ui) {
+        egui::Panel::top("top_panel").show(ui, |ui| {
+            egui::MenuBar::new().ui(ui, |ui| {
                 // NOTE: no File->Quit on web pages!
                 let is_web = cfg!(target_arch = "wasm32");
                 if !is_web {
@@ -91,13 +96,13 @@ impl SecCardGameApp {
                             self.last_event = Some(AppEvent::new_game());
                         }
                         if ui.button("Quit").clicked() {
-                            ctx.send_viewport_cmd(egui::ViewportCommand::Close);
+                            ui.ctx().send_viewport_cmd(egui::ViewportCommand::Close);
                         }
                     });
                     ui.add_space(16.0);
                 }
 
-                egui::widgets::global_dark_light_mode_buttons(ui);
+                egui::widgets::global_theme_preference_buttons(ui);
                 ui.add_space(16.0);
                 egui::gui_zoom::zoom_menu_buttons(ui);
             });