+++ gui/src/game_view/card_window/card_window.rs
@@ -3,7 +3,7 @@ use crate::game_view::card_window::card_view_model::{CardContent, CardMarker};
 use eframe::epaint::FontFamily;
 use egui::{Color32, Context, Frame, Label, Order, Pos2, RichText, Rounding, Style, TextBuffer, Ui, Vec2, WidgetText, Window};
 use game_lib::cards::properties::incident_impact::IncidentImpact;
-use rand::Rng;
+use rand::RngExt;
 
 pub struct CardWindow<'a> {
     max_size: Vec2,
@@ -11,7 +11,7 @@ pub struct CardWindow<'a> {
     content: &'a CardContent,
 }
 
-pub fn display_card<F>(card: &CardContent, command_callback: &mut F, ctx: &Context, ui: &mut Ui)
+pub fn display_card<F>(card: &CardContent, command_callback: &mut F, ui: &mut Ui)
 where
     F: FnMut(Command),
 {
@@ -21,7 +21,7 @@ where
         content: card,
     };
 
-    create_window(window, command_callback, ctx, ui)
+    create_window(window, command_callback, ui)
 }
 
 fn create_style(is_incident_target: bool, ui: &mut Ui) -> Style {
@@ -37,15 +37,15 @@ fn create_style(is_incident_target: bool, ui: &mut Ui) -> Style {
     style
 }
 
-fn create_window<F>(data: CardWindow, command_callback: &mut F, ctx: &Context, ui: &mut Ui)
+fn create_window<F>(data: CardWindow, command_callback: &mut F, ui: &mut Ui)
 where
     F: FnMut(Command),
 {
     let card = data.content;
     let area = ui.available_size();
-    let mut rng = rand::thread_rng();
-    let offset_x = rng.gen_range(-20.0..20.0);
-    let offset_y = rng.gen_range(-20.0..20.0);
+    let mut rng = rand::rng();
+    let offset_x = rng.random_range(-20.0..20.0);
+    let offset_y = rng.random_range(-20.0..20.0);
     let new_pos = Pos2::new(area.x / 3.0 + offset_x, area.y / 3.0 + offset_y);
     let style = create_style(data.content.is_incident_target, ui);
 
@@ -64,7 +64,7 @@ where
         generic_card_window
     };
 
-    customized_card_window.show(ctx, |ui| create_card_window(command_callback, card, ui));
+    customized_card_window.show(ui, |ui| create_card_window(command_callback, card, ui));
 }
 
 fn create_card_window<F>(cmd_callback: &mut F, card: &CardContent, ui: &mut Ui)